requirements.txt
attrs==22.2.0
et-xmlfile==1.1.0
iniconfig==1.1.1
numpy==1.23.5
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pluggy==1.0.0
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.6
# Editable install with no version control (rates==0.1.0)
-e /home/egor_usual/rates/src
six==1.16.0
